Update # P034-和差积商.py

+++ # P034-和差积商.py
@@ -1,6 +1,6 @@
 # 输入2个整数，求：和，差，积，商
 # 输入：2个整数（第二个数是非零整数）
 # 输出：4个整数，依次为和、差、积和商，数据间用空格分隔。
-a,b=map(float,input().split())
+a,b=map(int,input().split())
 
 print(a+b,a-b,a*b,a/b)
